Initial threaded callback support.
This feature is intended for use with HTTP queries and the like that may
take a little while to run.

Usage:

Execute self.bot.set_callback(
    callable (code you want to run in a thread),
    callable(result) (handles the result)
    args,
    kwargs,  (these two pass to the first callable)
    )

Any exceptions are returned as the result, so check the type before
acting on the result.

Haven't checked to see how reply_to and the like would factor in, but no
reason why you could define a function in local scope and set that as
the handler.